refactoring and mypy
