+++ .pre-commit-config.yaml
@@ -9,6 +9,7 @@ repos:
     rev: v5.8.0
     hooks:
       - id: isort
+        args: ["--profile", "black"]
   - repo: https://github.com/pre-commit/pre-commit-hooks
     rev: v3.4.0
     hooks:
